fix utf16 lossy conversion on odd length strings
If a stings has the 0xFEFF BOM marker, utf16be_to_char() will be called regardless
of its length.

If the string is not actually utf16 encoded, and has an odd length,
the parser will panic on the out of bound access caused by `w[1]` on
the remainder chunk of length 1.

Mitigate this by calling chunks_exact() instead. The last byte will be ignored
during the conversion.